requirements.txt
pandas
requests
Pillow
transformers>=4.45.1
matplotlib
openpyxl